[Feat] Add super class for unified classes
