+++ src/classes/Paragraph.py
@@ -1,11 +1,10 @@
 import re
 from dataclasses import dataclass, field
-
-from src.helpers.enums import AlignmentEnum
+from src.classes.superclass.Element import Element
 
 
 @dataclass
-class Paragraph:
+class Paragraph(Element):
     """
     Description: a inified class representing a text paragraph, its properties? styles and content
 
@@ -89,7 +88,8 @@ class Paragraph:
                 Calculates the type of the first character of a paragraph
 
     """
-    _line_spacing: float
+    _font_name: str
+    _text_size: float
     _text: str
     _count_of_sp_sbl: int = field(init=False)
     _count_sbl: int = field(init=False)
@@ -97,25 +97,12 @@ class Paragraph:
     _uppercase: bool = field(init=False)
     _last_sbl: str = field(init=False)
     _first_key: str = field(init=False)
-    _indent: float
-    _font_name: str
-    _text_size: float
-
-    _alignment: AlignmentEnum = None
-    _mrgrg: float = None
-    _mrglf: float = None
-    _mrgtop: float = None
-    _mrgbtm: float = None
     _bold: bool = None
     _italics: bool = None
     _underlining: bool = None
     _sub_text: bool = None
     _super_text: bool = None
     _color_text: str = None
-    _page_breake_before: bool = None
-    _keep_lines_together: bool = None
-    _keep_with_next: bool = None
-    _outline_level: str = None
     _no_change_fontname: bool = None
     _no_change_text_size: bool = None
 
@@ -239,70 +226,6 @@ class Paragraph:
         self._text = text
 
     @property
-    def keep_lines_together(self):
-        return self._keep_lines_together
-
-    @keep_lines_together.setter
-    def keep_lines_together(self, keep_lines_together):
-        self._keep_lines_together = keep_lines_together
-
-    @property
-    def outline_level(self):
-        return self._outline_level
-
-    @outline_level.setter
-    def outline_level(self, outline_level):
-        self._outline_level = outline_level
-
-    @property
-    def keep_with_next(self):
-        return self._keep_with_next
-
-    @keep_with_next.setter
-    def keep_with_next(self, keep_with_next):
-        self._keep_with_next = keep_with_next
-
-    @property
-    def indent(self):
-        return self._indent
-
-    @indent.setter
-    def indent(self, indent):
-        self._indent = indent
-
-    @property
-    def mrgrg(self):
-        return self._mrgrg
-
-    @mrgrg.setter
-    def mrgrg(self, mrgrg):
-        self._mrgrg = mrgrg
-
-    @property
-    def mrglf(self):
-        return self._mrglf
-
-    @mrglf.setter
-    def mrglf(self, mrglf):
-        self._mrglf = mrglf
-
-    @property
-    def mrgtop(self):
-        return self._mrgtop
-
-    @mrgtop.setter
-    def mrgtop(self, mrgtop):
-        self._mrgtop = mrgtop
-
-    @property
-    def mrgbtm(self):
-        return self._mrgbtm
-
-    @mrgbtm.setter
-    def mrgbtm(self, mrgbtm):
-        self._mrgbtm = mrgbtm
-
-    @property
     def font_name(self):
         return self._font_name
 
@@ -317,17 +240,6 @@ class Paragraph:
     @color_text.setter
     def color_text(self, color_text):
         self._color_text = color_text
-
-    @property
-    def line_spacing(self):
-        return self._line_spacing
-
-    @line_spacing.setter
-    def line_spacing(self, value):
-        if value >= 0:
-            self._line_spacing = value
-        else:
-            raise ValueError
 
     @property
     def bold(self):
